requirements.txt
# 移动端自动化测试依赖

# Android自动化（必需）
uiautomator2>=2.16.0
adbutils>=1.2.0

# 图片处理（视觉识别需要）
Pillow>=10.0.0

# 多模态AI支持（可选，用于视觉识别）
dashscope>=1.10.0  # 通义千问VL

# 环境变量配置（AI配置需要）
python-dotenv>=1.0.0

# H5/WebView支持（可选，用于复杂H5场景）
# 如果你的App包含复杂H5内容，取消下面的注释：
# Appium-Python-Client>=3.0.0
# selenium>=4.0.0

# iOS支持（可选，需要macOS和Xcode）
# 方案1：tidevice + facebook-wda（推荐，API类似uiautomator2）
tidevice>=0.11.0  # 阿里开源的iOS设备管理工具
facebook-wda>=1.4.0  # Facebook的iOS自动化库

# 方案2：Appium（传统方案，配置较复杂）
# Appium-Python-Client>=3.0.0

# 异步支持（Python标准库，无需安装）
# asyncio

# XML解析（Python标准库，无需安装）
# xml.etree.ElementTree

# 测试框架和报告（Allure支持）
pytest>=8.0.0
pytest-asyncio>=0.21.0
allure-pytest>=2.13.0

# MCP 协议支持（用于 Cursor AI 集成）
mcp>=0.9.0

# HTTP服务器支持（可选，用于局域网共享）
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

